+++ src/eas_3d_pattern/util_func/report.py
@@ -7,7 +7,7 @@ import pandas as pd
 from tqdm import tqdm
 
 from ..parser import AntennaPattern
-from ..sector_definitions import SectorDefinition
+from ..sector import Sector, from_preset
 
 logger = logging.getLogger(__name__)
 
@@ -117,14 +117,14 @@ def generate_report_eas(
 
 def _process_a_file(
     file_path: Path,
-) -> tuple[pd.DataFrame, AntennaPattern, SectorDefinition] | None:
+) -> tuple[pd.DataFrame, AntennaPattern, Sector] | None:
     """Process a file and return the processed data.
 
     Args:
         file_path (Path): The path to the file to be processed.
 
     Returns:
-        tuple[pd.DataFrame, AntennaPattern, SectorDefinition] | None:
+        tuple[pd.DataFrame, AntennaPattern, Sector] | None:
             A tuple containing the processed data, the AntennaPattern instance,
             and the SectorDefinition instance. If an exception is raised, None is
             returned.
@@ -136,7 +136,7 @@ def _process_a_file(
         pattern = AntennaPattern(str(file_path), validate=False)
         data = pattern.get_metadata_dict()
         top_border = pattern.calculate_top_3db_point(power=False)
-        eas_sectors = SectorDefinition(load_default=True, top_border=top_border)
+        eas_sectors = from_preset("eas", top_border=top_border)
         if (data["Phi_HPBW"] <= 50) & (pattern.pattern.peak_coordinates[1] < -20):
             logger.info(
                 "Reporting: Identified dual beam antenna. Overwriting sectors to dual beam definition for reporting."
@@ -197,7 +197,7 @@ def _process_a_file(
 
 def _save_figure(
     pattern: AntennaPattern,
-    sector_definitions: SectorDefinition,
+    sector_definitions: Sector,
     output_directory: Path,
     remove_layout_components: bool = False,
 ) -> None: